+++ src/pytest_check/check_functions.py
@@ -91,137 +91,150 @@ def assert_equal(a: object, b: object, msg: str = "") -> None:  # pragma: no cov
     assert a == b, msg
 
 
-def equal(a: object, b: object, msg: str = "") -> bool:
+def equal(a: object, b: object, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if a == b:
         return True
     else:
-        log_failure(f"check {a} == {b}", msg)
+        log_failure(f"check {a} == {b}", msg, xfail=xfail)
         return False
 
 
-def not_equal(a: object, b: object, msg: str = "") -> bool:
+def not_equal(a: object, b: object, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if a != b:
         return True
     else:
-        log_failure(f"check {a} != {b}", msg)
+        log_failure(f"check {a} != {b}", msg, xfail=xfail)
         return False
 
 
-def is_(a: object, b: object, msg: str = "") -> bool:
+def is_(a: object, b: object, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if a is b:
         return True
     else:
-        log_failure(f"check {a} is {b}", msg)
+        log_failure(f"check {a} is {b}", msg, xfail=xfail)
         return False
 
 
-def is_not(a: object, b: object, msg: str = "") -> bool:
+def is_not(a: object, b: object, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if a is not b:
         return True
     else:
-        log_failure(f"check {a} is not {b}", msg)
+        log_failure(f"check {a} is not {b}", msg, xfail=xfail)
         return False
 
 
-def is_true(x: object, msg: str = "") -> bool:
+def is_true(x: object, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if bool(x):
         return True
     else:
-        log_failure(f"check bool({x})", msg)
+        log_failure(f"check bool({x})", msg, xfail=xfail)
         return False
 
 
-def is_false(x: object, msg: str = "") -> bool:
+def is_false(x: object, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if not bool(x):
         return True
     else:
-        log_failure(f"check not bool({x})", msg)
+        log_failure(f"check not bool({x})", msg, xfail=xfail)
         return False
 
 
-def is_none(x: object, msg: str = "") -> bool:
+def is_none(x: object, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if x is None:
         return True
     else:
-        log_failure(f"check {x} is None", msg)
+        log_failure(f"check {x} is None", msg, xfail=xfail)
         return False
 
 
-def is_not_none(x: object, msg: str = "") -> bool:
+def is_not_none(x: object, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if x is not None:
         return True
     else:
-        log_failure(f"check {x} is not None", msg)
+        log_failure(f"check {x} is not None", msg, xfail=xfail)
         return False
 
 
-def is_nan(a: SupportsFloat | SupportsIndex, msg: str = "") -> bool:
+def is_nan(
+    a: SupportsFloat | SupportsIndex, msg: str = "", xfail: str | None = None
+) -> bool:
     __tracebackhide__ = True
     if math.isnan(a):
         return True
     else:
-        log_failure(f"check {a} is NaN", msg)
+        log_failure(f"check {a} is NaN", msg, xfail=xfail)
         return False
 
 
-def is_not_nan(a: SupportsFloat | SupportsIndex, msg: str = "") -> bool:
+def is_not_nan(
+    a: SupportsFloat | SupportsIndex, msg: str = "", xfail: str | None = None
+) -> bool:
     __tracebackhide__ = True
     if not math.isnan(a):
         return True
     else:
-        log_failure(f"check {a} is not NaN", msg)
+        log_failure(f"check {a} is not NaN", msg, xfail=xfail)
         return False
 
 
-def is_in(a: _T, b: Container[_T], msg: str = "") -> bool:
+def is_in(a: _T, b: Container[_T], msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if a in b:
         return True
     else:
-        log_failure(f"check {a} in {b}", msg)
+        log_failure(f"check {a} in {b}", msg, xfail=xfail)
         return False
 
 
-def is_not_in(a: _T, b: Container[_T], msg: str = "") -> bool:
+def is_not_in(a: _T, b: Container[_T], msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if a not in b:
         return True
     else:
-        log_failure(f"check {a} not in {b}", msg)
+        log_failure(f"check {a} not in {b}", msg, xfail=xfail)
         return False
 
 
 _TypeTuple = Union[type, tuple["_TypeTuple", ...]]
 
 
-def is_instance(a: object, b: _TypeTuple, msg: str = "") -> bool:
+def is_instance(
+    a: object, b: _TypeTuple, msg: str = "", xfail: str | None = None
+) -> bool:
     __tracebackhide__ = True
     if isinstance(a, b):
         return True
     else:
-        log_failure(f"check isinstance({a}, {b})", msg)
+        log_failure(f"check isinstance({a}, {b})", msg, xfail=xfail)
         return False
 
 
-def is_not_instance(a: object, b: _TypeTuple, msg: str = "") -> bool:
+def is_not_instance(
+    a: object, b: _TypeTuple, msg: str = "", xfail: str | None = None
+) -> bool:
     __tracebackhide__ = True
     if not isinstance(a, b):
         return True
     else:
-        log_failure(f"check not isinstance({a}, {b})", msg)
+        log_failure(f"check not isinstance({a}, {b})", msg, xfail=xfail)
         return False
 
 
 def almost_equal(
-    a: object, b: object, rel: Any = None, abs: Any = None, msg: str = ""
+    a: object,
+    b: object,
+    rel: Any = None,
+    abs: Any = None,
+    msg: str = "",
+    xfail: str | None = None,
 ) -> bool:
     """
     For rel and abs tolerance, see:
@@ -231,12 +244,19 @@ def almost_equal(
     if a == pytest.approx(b, rel, abs):
         return True
     else:
-        log_failure(f"check {a} == pytest.approx({b}, rel={rel}, abs={abs})", msg)
+        log_failure(
+            f"check {a} == pytest.approx({b}, rel={rel}, abs={abs})", msg, xfail=xfail
+        )
         return False
 
 
 def not_almost_equal(
-    a: object, b: object, rel: Any = None, abs: Any = None, msg: str = ""
+    a: object,
+    b: object,
+    rel: Any = None,
+    abs: Any = None,
+    msg: str = "",
+    xfail: str | None = None,
 ) -> bool:
     """
     For rel and abs tolerance, see:
@@ -246,121 +266,153 @@ def not_almost_equal(
     if a != pytest.approx(b, rel, abs):
         return True
     else:
-        log_failure(f"check {a} != pytest.approx({b}, rel={rel}, abs={abs})", msg)
+        log_failure(
+            f"check {a} != pytest.approx({b}, rel={rel}, abs={abs})", msg, xfail=xfail
+        )
         return False
 
 
 @overload
-def greater(a: float, b: float, msg: str = "") -> bool: ...
+def greater(a: float, b: float, msg: str = "", xfail: str | None = None) -> bool: ...
 
 
 @overload
 def greater(
-    a: SupportsFloat | SupportsIndex, b: SupportsFloat | SupportsIndex, msg: str = ""
+    a: SupportsFloat | SupportsIndex,
+    b: SupportsFloat | SupportsIndex,
+    msg: str = "",
+    xfail: str | None = None,
 ) -> bool: ...
 
 
 @overload
-def greater(a: _ComparableGreaterThan[_CmpT], b: _CmpT, msg: str = "") -> bool: ...
+def greater(
+    a: _ComparableGreaterThan[_CmpT], b: _CmpT, msg: str = "", xfail: str | None = None
+) -> bool: ...
 
 
-def greater(a: Any, b: Any, msg: str = "") -> bool:
+def greater(a: Any, b: Any, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if a is None or b is None:
-        log_failure(f"check {a} > {b}: None cannot be compared", msg)
+        log_failure(f"check {a} > {b}: None cannot be compared", msg, xfail=xfail)
         return False
     elif a > b:
         return True
     else:
-        log_failure(f"check {a} > {b}", msg)
+        log_failure(f"check {a} > {b}", msg, xfail=xfail)
         return False
 
 
 @overload
-def greater_equal(a: float, b: float, msg: str = "") -> bool: ...
-
-
-@overload
 def greater_equal(
-    a: SupportsFloat | SupportsIndex, b: SupportsFloat | SupportsIndex, msg: str = ""
+    a: float, b: float, msg: str = "", xfail: str | None = None
 ) -> bool: ...
 
 
 @overload
 def greater_equal(
-    a: _ComparableGreaterThanOrEqual[_CmpT], b: _CmpT, msg: str = ""
+    a: SupportsFloat | SupportsIndex,
+    b: SupportsFloat | SupportsIndex,
+    msg: str = "",
+    xfail: str | None = None,
 ) -> bool: ...
 
 
-def greater_equal(a: Any, b: Any, msg: str = "") -> bool:
+@overload
+def greater_equal(
+    a: _ComparableGreaterThanOrEqual[_CmpT],
+    b: _CmpT,
+    msg: str = "",
+    xfail: str | None = None,
+) -> bool: ...
+
+
+def greater_equal(a: Any, b: Any, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if a is None or b is None:
-        log_failure(f"check {a} >= {b}: None cannot be compared", msg)
+        log_failure(f"check {a} >= {b}: None cannot be compared", msg, xfail=xfail)
         return False
     elif a >= b:
         return True
     else:
-        log_failure(f"check {a} >= {b}", msg)
+        log_failure(f"check {a} >= {b}", msg, xfail=xfail)
         return False
 
 
 @overload
-def less(a: float, b: float, msg: str = "") -> bool: ...
+def less(a: float, b: float, msg: str = "", xfail: str | None = None) -> bool: ...
 
 
 @overload
 def less(
-    a: SupportsFloat | SupportsIndex, b: SupportsFloat | SupportsIndex, msg: str = ""
+    a: SupportsFloat | SupportsIndex,
+    b: SupportsFloat | SupportsIndex,
+    msg: str = "",
+    xfail: str | None = None,
 ) -> bool: ...
 
 
 @overload
-def less(a: _ComparableLessThan[_CmpT], b: _CmpT, msg: str = "") -> bool: ...
+def less(
+    a: _ComparableLessThan[_CmpT], b: _CmpT, msg: str = "", xfail: str | None = None
+) -> bool: ...
 
 
-def less(a: Any, b: Any, msg: str = "") -> bool:
+def less(a: Any, b: Any, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if a is None or b is None:
-        log_failure(f"check {a} < {b}: None cannot be compared", msg)
+        log_failure(f"check {a} < {b}: None cannot be compared", msg, xfail=xfail)
         return False
     elif a < b:
         return True
     else:
-        log_failure(f"check {a} < {b}", msg)
+        log_failure(f"check {a} < {b}", msg, xfail=xfail)
         return False
 
 
 @overload
-def less_equal(a: float, b: float, msg: str = "") -> bool: ...
+def less_equal(a: float, b: float, msg: str = "", xfail: str | None = None) -> bool: ...
 
 
 @overload
 def less_equal(
-    a: SupportsFloat | SupportsIndex, b: SupportsFloat | SupportsIndex, msg: str = ""
+    a: SupportsFloat | SupportsIndex,
+    b: SupportsFloat | SupportsIndex,
+    msg: str = "",
+    xfail: str | None = None,
 ) -> bool: ...
 
 
 @overload
 def less_equal(
-    a: _ComparableLessThanOrEqual[_CmpT], b: _CmpT, msg: str = ""
+    a: _ComparableLessThanOrEqual[_CmpT],
+    b: _CmpT,
+    msg: str = "",
+    xfail: str | None = None,
 ) -> bool: ...
 
 
-def less_equal(a: Any, b: Any, msg: str = "") -> bool:
+def less_equal(a: Any, b: Any, msg: str = "", xfail: str | None = None) -> bool:
     __tracebackhide__ = True
     if a is None or b is None:
-        log_failure(f"check {a} <= {b}: None cannot be compared", msg)
+        log_failure(f"check {a} <= {b}: None cannot be compared", msg, xfail=xfail)
         return False
     elif a <= b:
         return True
     else:
-        log_failure(f"check {a} <= {b}", msg)
+        log_failure(f"check {a} <= {b}", msg, xfail=xfail)
         return False
 
 
 @overload
 def between(
-    b: float, a: float, c: float, msg: str = "", ge: bool = False, le: bool = False
+    b: float,
+    a: float,
+    c: float,
+    msg: str = "",
+    ge: bool = False,
+    le: bool = False,
+    xfail: str | None = None,
 ) -> bool: ...
 
 
@@ -372,6 +424,7 @@ def between(
     msg: str = "",
     ge: bool = False,
     le: bool = False,
+    xfail: str | None = None,
 ) -> bool: ...
 
 
@@ -383,44 +436,55 @@ def between(
     msg: str = "",
     ge: bool = False,
     le: bool = False,
+    xfail: str | None = None,
 ) -> bool: ...
 
 
 def between(
-    b: Any, a: Any, c: Any, msg: str = "", ge: bool = False, le: bool = False
+    b: Any,
+    a: Any,
+    c: Any,
+    msg: str = "",
+    ge: bool = False,
+    le: bool = False,
+    xfail: str | None = None,
 ) -> bool:
     __tracebackhide__ = True
     if a is None or b is None or c is None:
-        log_failure(f"check {a} <= {b} <= {c}: None cannot be compared", msg)
+        log_failure(
+            f"check {a} <= {b} <= {c}: None cannot be compared", msg, xfail=xfail
+        )
         return False
     elif ge and le:
         if a <= b <= c:
             return True
         else:
-            log_failure(f"check {a} <= {b} <= {c}", msg)
+            log_failure(f"check {a} <= {b} <= {c}", msg, xfail=xfail)
             return False
     elif ge:
         if a <= b < c:
             return True
         else:
-            log_failure(f"check {a} <= {b} < {c}", msg)
+            log_failure(f"check {a} <= {b} < {c}", msg, xfail=xfail)
             return False
     elif le:
         if a < b <= c:
             return True
         else:
-            log_failure(f"check {a} < {b} <= {c}", msg)
+            log_failure(f"check {a} < {b} <= {c}", msg, xfail=xfail)
             return False
     else:
         if a < b < c:
             return True
         else:
-            log_failure(f"check {a} < {b} < {c}", msg)
+            log_failure(f"check {a} < {b} < {c}", msg, xfail=xfail)
             return False
 
 
 @overload
-def between_equal(b: float, a: float, c: float, msg: str = "") -> bool: ...
+def between_equal(
+    b: float, a: float, c: float, msg: str = "", xfail: str | None = None
+) -> bool: ...
 
 
 @overload
@@ -429,6 +493,7 @@ def between_equal(
     a: SupportsFloat | SupportsIndex,
     c: SupportsFloat | SupportsIndex,
     msg: str = "",
+    xfail: str | None = None,
 ) -> bool: ...
 
 
@@ -438,6 +503,7 @@ def between_equal(
     a: _CmpT,
     c: _CmpT,
     msg: str = "",
+    xfail: str | None = None,
 ) -> bool: ...
 
 
@@ -446,11 +512,12 @@ def between_equal(
     a: Any,
     c: Any,
     msg: str = "",
+    xfail: str | None = None,
 ) -> bool:
     __tracebackhide__ = True
-    return between(b, a, c, msg, ge=True, le=True)
+    return between(b, a, c, msg, ge=True, le=True, xfail=xfail)
 
 
-def fail(msg: str) -> None:
+def fail(msg: str, xfail: str | None = None) -> None:
     __tracebackhide__ = True
-    log_failure(msg)
+    log_failure(msg, xfail=xfail)